Refine precision in astrodynamics tests
Updated the ASSERT_NEAR assertions in astrodynamics tests by reducing the tolerance from 6 to 1E-06. Through this change, we ensure that tiny discrepancies, potentially resulting from floating-point arithmetic, in the expected and produced test results can be caught and addressed effectively. This contributes to better accuracy and robustness of the astrodynamic calculations in our application.