Add `CanExecute` function to ApogeeHeightChangingManeuver
This commit adds a `CanExecute` function to the `ApogeeHeightChangingManeuver` class. The purpose of this function is to evaluate whether or not a given maneuver can be executed based on the parameters of the orbital point. Explicit implementation of this function in this specific class makes the logic around the maneuver execution more explicit and easier to understand.