Update conditions for execution in maneuver classes
The CanExecute function in ApogeeHeightChangingManeuver, PhasingManeuver and PerigeeHeightChangingManeuver classes has been updated to account for circular orbits. Now, if the orbit is circular, the position and epoch of the maneuver point are updated. Previously, it only returned true without updating these parameters. Also, replaced hard-coded return false with a call to the base method, increasing code flexibility. In addition, the ManeuverPointComputation method in these classes has been revised to return appropriate vector instead of a zero vector for better consistency with the method semantics.