Feature/improvements (#148)
* Refine precision in astrodynamics tests

Updated the ASSERT_NEAR assertions in astrodynamics tests by reducing the tolerance from 6 to 1E-06. Through this change, we ensure that tiny discrepancies, potentially resulting from floating-point arithmetic, in the expected and produced test results can be caught and addressed effectively. This contributes to better accuracy and robustness of the astrodynamic calculations in our application.

* Improve maneuver execution process and update contact details

The maneuver execution process has been improved by adding new optional members (m_maneuverPointTarget, m_maneuverPointUpdate and m_isInbound) in 'ManeuverBase' class. This will ensure smooth execution of the maneuvers. In addition, new virtual methods 'UpdateManeuverPoint' and 'ManeuverPointComputation' have been added for more flexibility in maneuver execution. Changes were required to improve the system's performance and make the codebase more maintainable.

Also, the contact details of the author has been updated in 'TestManeuver.h' file.

Several other minor corrections and additions have been made throughout the codebase.

* Add compute maneuver point function

Remove bool::CanExecute as it's redundant and add function for computing the maneuver point. This computation is done using various factors such as current position, velocity, gravitational and perturbation forces. Update the tests to match these changes.

* Update orbit maneuvers and tests accuracy in Astrodynamics

This commit removes the unnecessary boolean function 'CanExecute' from the ManeuverBase class and adds a function to compute the maneuver point in the OrbitalPlaneChangingManeuver class. This maneuver point calculation is essential for orbital dynamics computations as it uses multiple variables such as position, velocity, and forces to provide a more accurate maneuvering action. Additionally, the test cases in the PlaneChangingManeuver and APITests have been updated to match these code modifications. The given expected results in the ASSERT statements have been modified to match the improved precision provided by these changes.

* Add `CanExecute` function to ApogeeHeightChangingManeuver

This commit adds a `CanExecute` function to the `ApogeeHeightChangingManeuver` class. The purpose of this function is to evaluate whether or not a given maneuver can be executed based on the parameters of the orbital point. Explicit implementation of this function in this specific class makes the logic around the maneuver execution more explicit and easier to understand.

* Update conditions for execution in maneuver classes

The CanExecute function in ApogeeHeightChangingManeuver, PhasingManeuver and PerigeeHeightChangingManeuver classes has been updated to account for circular orbits. Now, if the orbit is circular, the position and epoch of the maneuver point are updated. Previously, it only returned true without updating these parameters. Also, replaced hard-coded return false with a call to the base method, increasing code flexibility. In addition, the ManeuverPointComputation method in these classes has been revised to return appropriate vector instead of a zero vector for better consistency with the method semantics.

* Update maneuver execution conditions and computation

Added position and epoch updates to function CanExecute in ApogeeHeightChangingManeuver, PhasingManeuver, and PerigeeHeightChangingManeuver classes for circular orbits. Replaced hard-coded return false with a base method call and revised ManeuverPointComputation to return non-zero vectors. These changes made classes more flexible and consistent.

* fix pro

* fix combined maneuver

* Refactor CombinedManeuver and ManeuverBase computation

Removed redundant code fragments and adjusted calculations related to maneuvers in CombinedManeuver.cpp and ManeuverBase.cpp. The changes include updated methods for mean anomaly and apogee computation to make the code more efficient and precise. The 'if' condition checks related to orbital parameters were also revised for better accuracy. These changes, together with the updated test cases clearly improve the efficiency of the code and accuracy of calculations.

* Add method to update maneuver and improve calculations

Introduced a new method in ManeuverBase.h to update the maneuver point with new orbital parameters. This method will allow more flexibility when handling maneuvers and greatly reduces the likelihood of errors when updating maneuvers. Furthermore, abstract class for maneuver point computation was added to ensure the proper computation based on given orbital parameters. Overall, these changes significantly enhance the efficiency and accuracy of our astrodynamics computations.

* Refactor VectorTests to use new Planes test class

Updated VectorTests.cpp to use assertions against the newly created Planes.h test class, increasing modularity and ensuring testing is done against correct versions of plane calculations. This change refactors the tests to avoid direct dependencies on the Plane class by instead depending on a local test version, providing more flexibility for future updates and adjustments to the Plane class and vector test cases.